Revert "Revert v3.05 updates "merge Test websockets""
