+++ README.md
@@ -1,5 +1,35 @@
 # monopoli v3
-rework game monopoli menggunakan NextJS, Pubnub dan Tailwind
+rework game monopoli yang hasilnya ngebug doang 💀
+
+## Technologies
+- Next.js
+- TypeScript
+- Tailwind CSS
+- PubNub (websocket)
+- Redis (via Upstash)
+- Cloudinary (Avatar storage)
+- JWT Authentication
+
+## Features
+- Real-time multiplayer monopoly game
+- Multi-language support
+- Player chat system with emotes
+- Customizable game rooms:
+  - Different board types (normal/two-way)
+  - Various dice modes (1/2 dice)
+  - Multiple game modes (survive/laps)
+  - Adjustable starting money and loss conditions
+- Player profiles with:
+  - Customizable avatars
+  - Game statistics tracking
+  - Worst money lost tracking
+- Room list with:
+  - Player limits
+  - Password protection option
+  - Character selection
+- Interactive tutorial system
+- Sound effects
+- Mobile-responsive design
 
 #### Design UI
 *orang gk bisa design* 😎
@@ -11,3 +41,5 @@ rework game monopoli menggunakan NextJS, Pubnub dan Tailwind
 
 - game room
 <img src="https://github.com/aotti/monopoli-v3/blob/main/docs/design/monopoli-gameroom.jpg" width="70%" height="70%">
+
+readme by @Ni-zav
