+++ src/app/api/Controller.ts
@@ -77,9 +77,7 @@ export default class Controller {
 
     protected async redisReset(key: string) {
         // reset existing data
-        const resetResult = await redisClient.del(key)
-        console.log(resetResult);
-        
+        await redisClient.del(key)
     }
 
     protected filterPayload<T>(action: string, payload: T) {
@@ -135,10 +133,10 @@ export default class Controller {
         }
     }
 
-    protected async getOnlinePlayers(data: IPlayer, userAgent: string) {
+    protected async getOnlinePlayers(data: IPlayer, userAgent: string, action?: string) {
         const logData: Partial<ILoggedUsers> = {
             display_name: data.display_name,
-            status: 'online',
+            status: action?.match(/game|room join/) ? 'playing' : 'online',
             user_agent: userAgent
         }
         // renew my player
@@ -169,7 +167,6 @@ export default class Controller {
             }
             // check for account in use
             // match the user agent
-            console.log(action, 'checking account in use');
             const findUser = loggedPlayers.map(v => v.display_name).indexOf(payload.display_name)
             const isUserAgentMatch = matchUserAgent(payload.user_agent, findUser)
             if(findUser !== -1 && !isUserAgentMatch) {
@@ -209,7 +206,8 @@ export default class Controller {
             if(renewUser === -1) return this.respond(401, 'no access', [])
             // create token for timeout
             const timeoutToken = await this.generateAccessToken(payload as any, '5min')
-            // update my token
+            // update status & token
+            loggedPlayers[renewUser].status = payload.status
             loggedPlayers[renewUser].timeout_token = timeoutToken
             // save to redis
             await this.redisSet('loggedPlayers', loggedPlayers)
@@ -316,7 +314,7 @@ export default class Controller {
     }
 
     /**
-     * @param payload embedded access token to payload in route
+     * @param payload payload with access token key
      * @description verify access / refresh token
      * @returns verified token & get payload | error if no refresh token
      */
