+++ src/app/game/components/board/GameInfo.tsx
@@ -18,7 +18,9 @@ export default function GameInfo({ roomId }: {roomId: number}) {
             {/* mode */}
             <div className="flex flex-col">
                 <span> mode: </span>
-                <span className="text-green-400"> {gameState.gameRoomInfo[getGameRoomInfo]?.mode} </span>
+                <span className="text-green-400"> 
+                    {translateUI({lang: miscState.language, text: gameState.gameRoomInfo[getGameRoomInfo]?.mode as any})} 
+                </span>
             </div>
             {/* money lose */}
             <div className="flex flex-col">
