+++ src/config/change-logs.json
@@ -1,7 +1,7 @@
 {
     "updates": [
         {
-            "version": "v3.07",
+            "version": "v3.05",
             "date": "22 Apr 2025",
             "changes": [
                 "fix some bugs",
@@ -11,33 +11,19 @@
             ]
         },
         {
-            "version": "v3.06",
+            "version": "v3.04",
             "date": "13 Apr 2025",
             "changes": [
                 "fix some bugs"
             ]
         },
         {
-            "version": "v3.05",
+            "version": "v3.03",
             "date": "26 Mar 2025",
             "changes": [
                 "add new board",
                 "add notif on game start",
                 "fix some bugs"
-            ]
-        },
-        {
-            "version": "v3.04",
-            "date": "22 Mar 2025",
-            "changes": [
-                "add preparation before add new board"
-            ]
-        },
-        {
-            "version": "v3.03",
-            "date": "19 Mar 2025",
-            "changes": [
-                "fix screen bug"
             ]
         },
         {
